Okay main now uses bot
